requirements.txt
pypdf
sentence-transformers
faiss-cpu
ollama
numpy
streamlit
langchain-text-splitters